Move flavours/glitch/utils/emoji back to flavours/glitch/features/emoji
