+++ app/javascript/flavours/glitch/features/emoji/emoji_compressed.js
