+++ app/javascript/flavours/glitch/features/emoji/emoji_map.json
