+++ app/javascript/flavours/glitch/features/emoji/emoji_mart_data_light.js
