+++ app/javascript/flavours/glitch/features/emoji/emoji_mart_search_light.js
