+++ app/javascript/flavours/glitch/features/emoji/emoji_unicode_mapping_light.js
