+++ app/javascript/flavours/glitch/features/emoji/emoji_utils.js
