+++ app/javascript/flavours/glitch/features/getting_started/components/announcements.js
@@ -9,7 +9,7 @@ import { defineMessages, injectIntl, FormattedMessage, FormattedDate } from 'rea
 import { autoPlayGif, reduceMotion, disableSwiping } from 'flavours/glitch/initial_state';
 import elephantUIPlane from 'mastodon/../images/elephant_ui_plane.svg';
 import { mascot } from 'flavours/glitch/initial_state';
-import unicodeMapping from 'flavours/glitch/utils/emoji/emoji_unicode_mapping_light';
+import unicodeMapping from 'flavours/glitch/features/emoji/emoji_unicode_mapping_light';
 import classNames from 'classnames';
 import EmojiPickerDropdown from 'flavours/glitch/features/emoji_picker';
 import AnimatedNumber from 'flavours/glitch/components/animated_number';
