+++ app/javascript/flavours/glitch/reducers/custom_emojis.js
@@ -1,7 +1,7 @@
 import { List as ImmutableList, fromJS as ConvertToImmutable } from 'immutable';
 import { CUSTOM_EMOJIS_FETCH_SUCCESS } from 'flavours/glitch/actions/custom_emojis';
-import { search as emojiSearch } from 'flavours/glitch/utils/emoji/emoji_mart_search_light';
-import { buildCustomEmojis } from 'flavours/glitch/utils/emoji';
+import { search as emojiSearch } from 'flavours/glitch/features/emoji/emoji_mart_search_light';
+import { buildCustomEmojis } from 'flavours/glitch/features/emoji/emoji';
 
 const initialState = ImmutableList([]);
